Enable caching on all routes.
Uses Redis with defaults, so that needs to be installed and running.

Can't get STUPID FUCKING cache failover working in the
/src/utils/good-guy-cache.js file, don't know what's wrong with me.

Removed some unneeded code.

Improved some try/catch blocks and Boom error handling.

Turned off Hapi caching for the moment, we might not need it now that
good-guy-http is aggressively caching.