census.js
Temporary ( 😆 ) port of census api into the collator api.

With the amalgamation of departments there is constant changes to the structure a simpler solution, compared to the relational version, using rethinkdb by storing employees and all related information as a single entity (or "document"). This allows the information to change freely without the need for complex relations. However, this comes with a few downsides such as not being able to infer an organisation structure from the relations, as they don't exist anymore!

All endpoints and functionally are the same as the previous API so minimal changes are needed to adapt this new API but with the schema being different templates and some code will need to be refactored to reflect this.

Approved-by: Drew Robinson <drew.robinson@gmail.com>